test: reuse same test
